refactor(#2562 P5): split discord.rs into per-concern submodules
Pure structural refactor, zero behavior change. discord.rs had grown to
2482 LOC (P0-P4 completion work); splits it into src/channel/discord/,
mirroring channel::telegram's module layout (mod.rs declares submodules
+ pub(crate) use glob re-exports, no logic in mod.rs itself):

- state.rs      — DiscordState
- adapter.rs    — DiscordBindingPayload, DiscordChannel + impl, discord_caps,
                   impl Channel for DiscordChannel (mirrors telegram/adapter.rs
                   bundling BindingPayload + Channel struct + trait impl)
- protocol.rs   — gateway wire-format mapping (GatewayFrame, opcode/HELLO/
                   IDENTIFY parsing, ChannelEvent mapping)
- gateway.rs    — connection lifecycle (start_gateway, restart supervisor,
                   backoff, GATEWAY_DEAD)
- bootstrap.rs  — build_http_client, init_from_config
- inbound.rs    — dispatch_channel_event, spawn_inbound_dispatcher
- send.rs       — outbound message body construction
- keepalive.rs  — auto-archive keepalive loop
- tests.rs      — the single existing mod-tests block moved verbatim
                   (same pattern as worktree_pool.rs -> worktree_pool/tests.rs;
                   exempt from the anti-monolith LOC ceiling via is_test_file)

Extraction was mechanical: sed line-range cuts from the original file's
own existing `// ---- Section ----` comment boundaries (already
well-factored), verified contiguous (no line dropped/duplicated) before
writing. The only non-mechanical edits: 8 items (discord_runtime,
block_on_value, DiscordChannel::state field, build_http_client,
GATEWAY_DEAD, GATEWAY_RESTART_MAX_ATTEMPTS, discord_gateway_backoff_delay,
should_restart_gateway) bumped from private to pub(super) so sibling
modules and tests.rs can reach them through mod.rs's re-exports —
mirrors telegram/state.rs's identical pub(super) block_on_value pattern.
19 include_str! fixture paths adjusted for the new nesting depth
(../../ -> ../../../, tests.rs moved one directory deeper).

Verification:
- Test-name-set diff (the #2600 review standard): 46 #[test]/#[tokio::test]
  functions before, 46 after, zero difference.
- Whitespace-normalized content diff (original file vs. split files
  reassembled in original section order): zero non-boilerplate lines
  differ — only added mod declarations/imports/pub(super) keywords/
  fixture-path depth, no logic changed.
- cargo test --features discord: 45 passed, 1 ignored (unchanged).
- cargo test --workspace --tests --features tray: identical 262
  pre-existing unrelated failures vs. the established baseline, 0 new.
- cargo build --features tray (discord feature OFF, matches Coverage/
  Check CI jobs): clean.
- cargo fmt --check, cargo clippy --all-targets --features discord
  -- -D warnings: clean.
- LOC invariant (tests/src_file_size_invariant.rs): passes; discord.rs
  was never in GRANDFATHERED (no entry to remove); every new file is
  well under the 2500 MAX_LOC ceiling (largest is adapter.rs at 508).

Refs #2562, task t-20260704130236500715-14440-5

Agend-Agent: gapfix-dev3
Agend-Task: t-20260704130236500715-14440-5
Agend-Branch: refactor/2562-p5-discord-split
